requirements.txt

appdirs==1.4.4
asgiref==3.2.7
certifi==2020.4.5.1
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
djangorestframework==3.11.0
dnspython==1.16.0
ez-setup==0.9
filelock==3.0.12
gunicorn==20.0.4
jedi==0.17.0
parso==0.7.0
pipenv==2018.11.26
pluggy==0.13.1
postgres==3.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pyls==0.1.6
python-jsonrpc-server==0.3.4
python-language-server==0.32.0
pytz==2020.1
pyxdg==0.26
requests==2.24.0
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.20
virtualenv-clone==0.5.4
whitenoise==5.0.1
